feat: migrate storybook to TS
