+++ package.json
@@ -76,13 +76,13 @@
     "@babel/parser": "7.21.8",
     "@chakra-ui/cli": "2.4.1",
     "@next/eslint-plugin-next": "13.4.3",
-    "@storybook/addon-actions": "7.0.10",
-    "@storybook/addon-essentials": "7.0.10",
-    "@storybook/addon-links": "7.0.10",
+    "@storybook/addon-actions": "7.0.18",
+    "@storybook/addon-essentials": "7.0.18",
+    "@storybook/addon-links": "7.0.18",
     "@storybook/addon-postcss": "2.0.0",
-    "@storybook/nextjs": "7.0.10",
-    "@storybook/react": "7.0.10",
-    "@storybook/theming": "7.0.10",
+    "@storybook/nextjs": "7.0.18",
+    "@storybook/react": "7.0.18",
+    "@storybook/theming": "7.0.18",
     "@svgr/cli": "8.0.1",
     "@trivago/prettier-plugin-sort-imports": "4.1.1",
     "@types/bcrypt": "5.0.0",
@@ -103,7 +103,7 @@
     "lint-staged": "13.2.2",
     "prettier": "2.8.8",
     "prisma": "4.14.0",
-    "storybook": "7.0.10",
+    "storybook": "7.0.18",
     "storybook-dark-mode": "3.0.0",
     "tsx": "3.12.7",
     "typescript": "5.0.4"
